[refactor] extract ldapjs-promise.js
